requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.4
django-filter==24.2
djangorestframework==3.15.1
frozendict==2.3.10
gunicorn==21.2.0
idna==3.7
Markdown==3.6
moralis==0.1.45
packaging==24.0
psycopg==3.1.18
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-dotenv==1.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.3.0
urllib3==1.26.18
